clean up upf-u code: move trtcm to a seperate file
