+++ 5gc/upf_u/meson.build
@@ -3,6 +3,7 @@ sources = files(
     'upf_u_config.c',
     'upf_u_arp.c',
     'upf_u_icmp.c',
+    'upf_u_trtcm.c',
     'upf_u_helper.c'
 )
