+++ 5gc/upf_u/upf_u_trtcm.c
@@ -1,0 +1,383 @@
+/*
+# Copyright 2026 University of California, Riverside and National Yang Ming Chiao Tung University
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# SPDX-License-Identifier: Apache-2.0
+*/
+
+#include "upf_context.h"
+#include "upf_u_trtcm.h"
+#include "utlt_debug.h"
+#include "upf_u_config.h"
+
+#include "../classifiers/classifier_wrapper.h"
+
+#define MIN(x, y) (((x) < (y)) ? (x) : (y))
+
+flow_entry_t iPFlows[APP_FLOWS_MAX];
+uint32_t iPFlowsLen = 0;
+uint32_t trTCMidx = 0;
+
+struct rte_meter_trtcm_profile app_trtcm_profile;
+struct rte_meter_trtcm_profile app_flow_trtcm_profile;
+struct rte_meter_trtcm app_flows[APP_FLOWS_MAX];
+
+static struct ue_hash_entry ue_hash[MAX_UE];
+struct ue_tb ue_table[MAX_UE];
+
+/* trTCM */
+struct rte_meter_trtcm_params app_trtcm_params = {
+	.cir = 125000,    // bytes per secs
+	.pir = 625000,    // bytes per secs
+	.cbs = 2048,
+	.pbs = 2048
+};
+
+int
+trtcmConfigFlowTables(void) {
+    uint32_t i;
+    int rtn;
+    if (likely(app_flows[0].tc > 0))
+        return 0;
+    
+    // config trtcm profile
+    rtn = rte_meter_trtcm_profile_config(&app_trtcm_profile,
+		&app_trtcm_params);
+	if (rtn)
+		return rtn;
+        
+    // config flow meters with trtcm profiles
+    for (i = 0; i < APP_FLOWS_MAX; i++){
+        rtn = rte_meter_trtcm_config(&app_flows[i], &app_trtcm_profile);
+        if (rtn)
+            return rtn;
+    }
+
+    UTLT_Info("Flow table configured.");
+    return 0;
+}
+
+int
+trtcmColorHandle(uint32_t pkt_len, uint64_t time, uint8_t qfi, struct rte_meter_trtcm_profile *target_profile) {
+    uint8_t out_color = 0;
+    // check configured flow
+    if (unlikely(app_trtcm_profile.cir_period == 0)) {
+        UTLT_Info("flow cir_period set err");
+        return -1;
+    }
+    if (unlikely(app_trtcm_profile.pir_period == 0)) {
+        UTLT_Info("flow pir_period set err");    
+        return -1;
+    }
+    out_color = (uint8_t) rte_meter_trtcm_color_blind_check(&app_flows[qfi], 
+        target_profile, 
+        time, 
+        pkt_len);
+    return out_color;
+}
+
+int
+trtcmPolicer(struct onvm_pkt_meta *meta, int color_result) {
+    if (meta->action == ONVM_NF_ACTION_DROP) {
+        meta->flags = RTE_COLOR_RED;
+        UTLT_Info("TB not enough & traffic flow");
+        return 0;
+    }
+    switch (color_result) {
+        case RTE_COLOR_RED:
+            UTLT_Info("\033[0;31mRED(%d)\033[0m, drop pkt", RTE_COLOR_RED);
+            meta->flags = RTE_COLOR_RED;
+            meta->action = ONVM_NF_ACTION_DROP;
+            break;
+        case RTE_COLOR_YELLOW:
+            UTLT_Info("\033[0;32mYELLOW(%d)\033[0m, best effort pkt fwd", RTE_COLOR_YELLOW);
+            meta->flags = RTE_COLOR_YELLOW;
+            meta->action = ONVM_NF_ACTION_DROP;
+            break;
+        case RTE_COLOR_GREEN:
+            UTLT_Info("\033[0;33mGREEEN(%d)\033[0m, guaranted pkt fwd.", RTE_COLOR_GREEN);
+            meta->flags = RTE_COLOR_GREEN;
+            meta->action = ONVM_NF_ACTION_OUT;
+            break;
+        default:
+            UTLT_Error("Unexpected trTCM color output.");
+            return 1;
+    }
+    return 0;
+}
+
+// static inline source_interface_t
+// PortToSourceInterface(uint16_t port) {
+//     if (port == g_n3_port)  return SRC_IF_ACCESS;
+//     if (port == g_n6_port)    return SRC_IF_CORE;
+//     if (port == g_sgi_port)     return SRC_IF_SGI_LAN;
+//     UTLT_Warning("PortToSourceInterface: unknown port %" PRIu16
+//              " (N3=%" PRIu16 " N6=%" PRIu16 " SGI=%" PRIu16
+//              ") — defaulting to ACCESS",
+//              port, g_n3_port, g_n6_port, g_sgi_port);
+
+//     return SRC_IF_ACCESS;
+// }
+
+static inline uint16_t
+SourceInterfaceToPort(source_interface_t srcIf) {
+    switch (srcIf) {
+      case SRC_IF_ACCESS:   return g_n3_port;
+      case SRC_IF_CORE:     return g_n6_port;
+      case SRC_IF_SGI_LAN:  return g_sgi_port;
+      case SRC_IF_CP_FUNC:
+      case SRC_IF_LI_FUNC:
+      default:
+        return -1; // Invalid/unsupported source interface
+    }
+}
+
+static inline int
+hashFunc(uint32_t subnet) {
+    return subnet % APP_FLOWS_MAX;
+}
+
+int
+ftSearch(uint32_t subnet) {
+    int index = hashFunc(subnet);
+    int original_index = index;
+
+    while (iPFlows[index].in_use) {
+        if (iPFlows[index].subnet == subnet) {
+            return iPFlows[index].flow_idx;
+        }
+        index = (index + 1) % APP_FLOWS_MAX;  // Linear Probing
+        
+        if (index == original_index) {
+            break;
+        }
+    }
+
+    return -1;  // Not found
+}
+
+static inline bool
+ftAddEntry(uint32_t subnet, int flow_idx) {
+    if (iPFlowsLen >= APP_FLOWS_MAX) {
+        printf("Error: Maximum flow entries reached.\n");
+        return false;
+    }
+
+    if (ftSearch(subnet) != -1) {
+        printf("Error: Subnet %u already exists.\n", subnet);
+        return false;
+    }
+
+    int index = hashFunc(subnet);
+    while (iPFlows[index].in_use) {         // Linear Probing
+        index = (index + 1) % APP_FLOWS_MAX;
+    }
+
+    // Insert the entry
+    iPFlows[index].subnet = subnet;
+    iPFlows[index].flow_idx = flow_idx;
+    iPFlows[index].in_use = true;
+    iPFlowsLen++;
+
+    return true;
+}
+
+void 
+initUeTable() {
+    for (int i = 0; i < MAX_UE; i++) {
+        ue_table[i].ue_ip = 0;
+        ue_table[i].ue_ambr = 0;
+        ue_table[i].ue_gbr = 0;
+        ue_table[i].ue_mbr = 0;
+
+        ue_table[i].ue_nqos_tb_params.tb_rate = 0;
+        ue_table[i].ue_nqos_tb_params.tb_depth = 0;
+        ue_table[i].ue_nqos_tb_params.tb_tokens = 0;
+        ue_table[i].ue_nqos_tb_params.last_cycle = rte_get_tsc_cycles();
+        ue_table[i].ue_nqos_tb_params.cur_cycles = rte_get_tsc_cycles();
+
+        ue_table[i].ue_qos_tb_params.tb_rate = 0;
+        ue_table[i].ue_qos_tb_params.tb_depth = 0;
+        ue_table[i].ue_qos_tb_params.tb_tokens = 0;
+        ue_table[i].ue_qos_tb_params.last_cycle = rte_get_tsc_cycles();
+        ue_table[i].ue_qos_tb_params.cur_cycles = rte_get_tsc_cycles();
+    }
+}
+
+static inline int
+ueHashFunc(uint32_t ip) { return ip % MAX_UE; }
+
+void
+ueHashInit(void) {
+    for (int i = 0; i < MAX_UE; i++)
+        ue_hash[i].in_use = false;
+}
+
+void
+ConfigureQerFlows(const UPDK_PDR *pdr, bool is_uplink) {
+    /* pdr->qer is the QFI-bearing (per-flow) QER selected by CP.
+     * It carries the correct MBR/GBR for flow-level trTCM metering. */
+    const UPDK_QER *qer = pdr ? pdr->qer : NULL;
+    if (!qer || !qer->flags.maximumBitrate) return;
+
+    bool has_fd = pdr->has_fd;
+
+    /* Use the same key that the DL policing path uses for ftSearch().
+     * pdr->meter_key was precomputed by UPF-C from the same port mapping,
+     * so insert and lookup are always consistent. */
+    uint32_t key = has_fd ? pdr->meter_key
+                          : (uint32_t)SourceInterfaceToPort(pdr->pdi.sourceInterface);
+
+    /* Only add on first miss — subsequent packets for the same key are a no-op */
+    if (ftSearch(key) >= 0) return;
+
+    UTLT_Info("QER ID: %u key: %u", qer->qerId, key);
+
+    struct rte_meter_trtcm_params trtcm_params = app_trtcm_params;
+
+    uint32_t mbr = is_uplink ? qer->maximumBitrate.ul : qer->maximumBitrate.dl;
+    trtcm_params.pir = mbr * 1000 / 8;
+
+    if (qer->flags.guaranteedBitrate) {
+        uint32_t gbr = is_uplink ? qer->guaranteedBitrate.ul : qer->guaranteedBitrate.dl;
+        trtcm_params.cir = gbr * 1000 / 8;
+    } else {
+        trtcm_params.cir = is_uplink ? 0 : 1;
+    }
+
+    if (!ftAddEntry(key, trTCMidx)) {
+        UTLT_Warning("FT add failed");
+    }
+    UTLT_Info("Successfully add %u(%d) %u", key, hashFunc(key), trTCMidx);
+
+    // Match config profile to what color-check later uses:
+    // DL + SDF present → app_flow_trtcm_profile; else app_trtcm_profile
+    if (!is_uplink && has_fd) {
+        rte_meter_trtcm_profile_config(&app_flow_trtcm_profile, &trtcm_params);
+        rte_meter_trtcm_config(&app_flows[trTCMidx], &app_flow_trtcm_profile);
+    } else {
+        rte_meter_trtcm_profile_config(&app_trtcm_profile, &trtcm_params);
+        rte_meter_trtcm_config(&app_flows[trTCMidx], &app_trtcm_profile);
+    }
+
+    if (is_uplink) {
+        UTLT_Info("Find MBR (UL: %lu) in QERs", qer->maximumBitrate.ul);
+        if (qer->flags.guaranteedBitrate)
+            UTLT_Info("Find GBR (UL: %lu) in QERs", qer->guaranteedBitrate.ul);
+    } else {
+        UTLT_Info("Find MBR (DL: %lu) in QERs", qer->maximumBitrate.dl);
+        if (qer->flags.guaranteedBitrate)
+            UTLT_Info("Find GBR (DL: %lu) in QERs", qer->guaranteedBitrate.dl);
+    }
+
+    UTLT_Info("TRTCM params: %d %d %d %d\n",
+              trtcm_params.cir, trtcm_params.pir,
+              trtcm_params.cbs, trtcm_params.pbs);
+
+    trTCMidx++;
+}
+
+/* Return ue_table index, or -1 */
+static inline int
+ueHashSearch(uint32_t ue_ip) {
+    int idx = ueHashFunc(ue_ip);
+    int start = idx;
+    while (ue_hash[idx].in_use) {
+        if (ue_hash[idx].ue_ip == ue_ip)
+            return ue_hash[idx].ue_idx;
+        idx = (idx + 1) % MAX_UE;
+        if (idx == start) break;
+    }
+    return -1;
+}
+
+static inline bool
+ueHashInsert(uint32_t ue_ip, int ue_idx) {
+    if (ueHashSearch(ue_ip) >= 0) return false; /* already present */
+    int idx = ueHashFunc(ue_ip);
+    while (ue_hash[idx].in_use)
+        idx = (idx + 1) % MAX_UE;
+    ue_hash[idx].ue_ip  = ue_ip;
+    ue_hash[idx].ue_idx = ue_idx;
+    ue_hash[idx].in_use = true;
+    return true;
+}
+
+/* Legacy wrapper — now O(1) via hash */
+int 
+findIndexByUeIpAddress(uint32_t ue_ip) {
+    return ueHashSearch(ue_ip);
+}
+
+int
+addEntrybyUeIp(uint32_t ue_ip, uint32_t ue_ambr, uint32_t ue_gbr, uint32_t ue_mbr) {
+    for (int i = 0; i < MAX_UE; i++) {
+        if (ue_table[i].ue_ip == 0) { // find unused
+            ue_table[i].ue_ip = ue_ip;
+            ue_table[i].ue_ambr = ue_ambr;
+            ue_table[i].ue_gbr = ue_gbr;
+
+            uint32_t qos_rate = MIN((ue_gbr + (ue_ambr - ue_gbr) / 2), ue_mbr);
+
+            ue_table[i].ue_qos_tb_params.tb_rate = qos_rate / 1000;
+            ue_table[i].ue_qos_tb_params.tb_depth = qos_rate;
+            ue_table[i].ue_qos_tb_params.tb_tokens = qos_rate;
+            ue_table[i].ue_qos_tb_params.last_cycle = rte_get_tsc_cycles();
+            ue_table[i].ue_qos_tb_params.cur_cycles = rte_get_tsc_cycles();
+            UTLT_Info("QoS Rate: %d", qos_rate);
+
+            uint32_t nqos_rate = ue_ambr - qos_rate;
+
+            ue_table[i].ue_nqos_tb_params.tb_rate = nqos_rate / 1000;
+            ue_table[i].ue_nqos_tb_params.tb_depth = nqos_rate;
+            ue_table[i].ue_nqos_tb_params.tb_tokens = nqos_rate;
+            ue_table[i].ue_nqos_tb_params.last_cycle = rte_get_tsc_cycles();
+            ue_table[i].ue_nqos_tb_params.cur_cycles = rte_get_tsc_cycles();
+            UTLT_Info("non QoS Rate: %d", nqos_rate);
+
+            ueHashInsert(ue_ip, i);
+            return i;  // Return the allocated index
+        }
+    }
+    return -1;  // Table full
+}
+
+void 
+updateTokenbyIndex(int index) {
+    if (index > -1) {
+        uint64_t cur_cycles;
+        uint64_t elapsed_cycles;
+        uint64_t tokens_produced;
+
+        cur_cycles = rte_get_tsc_cycles();
+        elapsed_cycles = cur_cycles - ue_table[index].ue_nqos_tb_params.last_cycle;
+
+        tokens_produced = (elapsed_cycles * ue_table[index].ue_nqos_tb_params.tb_rate * 125000) / rte_get_tsc_hz();
+        ue_table[index].ue_nqos_tb_params.tb_tokens += tokens_produced;
+        if (ue_table[index].ue_nqos_tb_params.tb_tokens > ue_table[index].ue_nqos_tb_params.tb_depth)
+            ue_table[index].ue_nqos_tb_params.tb_tokens = ue_table[index].ue_nqos_tb_params.tb_depth;
+        ue_table[index].ue_nqos_tb_params.last_cycle = cur_cycles;
+
+        elapsed_cycles = cur_cycles - ue_table[index].ue_qos_tb_params.last_cycle;
+        tokens_produced = (elapsed_cycles * ue_table[index].ue_qos_tb_params.tb_rate * 125000) / rte_get_tsc_hz();
+        ue_table[index].ue_qos_tb_params.tb_tokens += tokens_produced;
+        if (ue_table[index].ue_qos_tb_params.tb_tokens > ue_table[index].ue_qos_tb_params.tb_depth)
+            ue_table[index].ue_qos_tb_params.tb_tokens = ue_table[index].ue_qos_tb_params.tb_depth;
+        ue_table[index].ue_qos_tb_params.last_cycle = cur_cycles;
+
+        return;
+    }
+    UTLT_Error("UE IP not found in the table");
+    return;
+}
